udev: whitelist mlx4_core locally-administered MAC addresses in the persistent rule generator
IBM stated MAC addresses used by mlx4_core are stable even if they're
locally administrated.

So use them in this case since the generator is not able to deal with
cards having multiple ports.

Note: the virtual functions of this card don't work except for SLE15
on s390x anyway as it uses a special kernel.

[fbui: fixes bsc#1083158]