compat-rules: generate more compat by-id symlinks for NVMe devices
scsi_id acted differently depending on the NVMe revision implemented
by a device.

The previous rule only generated symlinks for the case, when a device
only reported it's vendor, device serial number and namespace id as an
identification string.

However depending on the NVMe revision implemented by the device, more
information might be available (eui, nguid, ...). In this case, those
were prefered by scsi_id to generate ID_SERIAL value and therefore the
symlink names.

If a system has been installed with an old version using scsi_id and
then upgraded to a version which generated the serial value via the
sysfs attributes, this could lead to wrong symlinks.

This patch also generates symlinks for cases where a device did not
returns the serial number but either a EUI64 or NGUID based device
identification.

Furthermore when the device firmware is updated, a new revision of
NVMe might be implemented and new attributes might suddenly become
available. To make sure that the new information available won't
prevent the generation of the "old" ID_SERIAL value, all symlinks are
generated now.

[fbui: fixes bsc#1095096]