test-watchdog: include pre-timeout in timeout setting
from bsc#1042933 comment#24

Documentation/watchdog/watchdog-api.txt:

"Some watchdog timers can be set to have a trigger go off before the
actual time they will reset the system.  This can be done with an NMI,
interrupt, or other mechanism.  This allows Linux to record useful
information (like panic information and kernel coredumps) before it
resets."

Hpwdt implements the pretimeout feature. For Proliant systems the length of
time for the pre-timeout is 9 seconds.  That is 9 seconds before the ASR would
reset the system, an NMI is sent to it.  Receipt of the NMI allows hpwdt to
initiate a crash dump.